Suppress stdout during Raveler post-processing.
Also, fix one last run-contours-std.py-related typo.

Phil and Don's Raveler post-processing scripts (create tiles, create
bounding boxes, compile stack to HDF5, and create contours) produce
boatloads of stdout printing of dubious use when not debugging. This
is now redirected to a temporary file, the contents of which are
presently ignored, but could be used in the future.