Major bug fix: peek only at valid queue items
MergeQueue.peek() was designed to look at the next item in the merge
queue without popping it. However, the next item *could* be invalid,
and cause a merge of a valid item past threshold!

This commit fixes that by having peek() discard invalid items until a
valid one is at the front of the queue.