Fix handling of boundary body in merge/learn
I noticed that the edge to the boundary body was not being preserved
during some merges. I rectified it by allowing edge merges even when
the boundary body is involved.

This caused problems when learning, because the boundary body has value
"inf" as its probability. Therefore, we now check during learning that
the boundary body is not being used.